move kinesis record to test suite data
