+++ example/spec/kinesisTests/KinesisTestErrorSpec.js
@@ -21,7 +21,7 @@ const {
 
 const { loadConfig } = require('../helpers/testUtils');
 
-const record = require('../../data/records/L2_HR_PIXC_product_0001-of-4154.json');
+const record = require('./data/records/L2_HR_PIXC_product_0001-of-4154.json');
 
 describe('The kinesisConsumer receives a bad record.', () => {
   const testRecordIdentifier = randomString();
