+++ example/spec/kinesisTests/data/records/L2_HR_PIXC_product_0001-of-4154.json
